Added in interpolation routines and change the halodata into a vector
With this update OrbWeaver can now interpolate the positions of the
orbiting halos even if they are missing for multiple snapshots. A cubic
spline is used to interpolate the halos position and velocities, for
other properties (Mvir,Rvir,Rmax,Vmax) a log interpolation is used.

In addition the halodata is now a vector so the interpolated halos can
be easily appended to the dataset.