Updated the OrbitData structure
The OrbitData stucture has been changed so instead of storing information
for every halo it only stores the data for interesting points in a halos
orbit. The data will initally only stored for crossing of 3,2,1 times the
host's rvir and peri-centric and apo-centric passages, to reduced the
memory footprint and the size of the outputted data.