Added calculations at peri/apo-centers
The code now calculates orbital properties at peri/apo-centers. Previously
the peri/apo-centers were when a halo had a sign change in its velocity
vector which meant a halo had many peri/apo-centers. Now this is only done
when a halo goes past its first pericenter (has a sign change in its velocity
vector within the radius of its host). This is to make sure the number of
orbits is accurately calculated.