Bug fix and update to way objects are appended to the orbitdata
There was a bug in the radial velocity that had a * instead of a + and
updated the first passage to be 0.5 an orbit. Also updated the way the
orbitdata is appended by instead of appended it is done instead after
a whole halo branch is done. This is done so further interpolation can
be done later for the passage points.