Added to output and updated functionality
Calculations of rperi and rapo have been added to the output and the eccentricity from the ratio of the peri/apo-centric distances found in the simulation. In addition the setting of the type of passage has been moved to after ther interpolation of the passage points. This is to make sure make sure these are correct since they may change after interpolation.