Various bug fixes and update to when a object is found to be in orbit.
Fixed a bug where the radial velocity was used instead of the relative velocity and corrected the eccentricity calculation to use the peri/apo-centric points. In addition previously a orbit was started if there was a passage point within one Rvir of its host, but this has now been updated to 3 Rvir of its host halo.