Calculating average angular momentum and energy since last passage now
Since calculaing the eccentricity from Peri-centric passage can lead to unbound (+ve) energies partitulary at the first few passages the average energy and angular momentum are calculated. So this eccentricity is more consistent with the one calculated from rperi and rapo. To also remove unphysical orbits a cut is made where rapo-rper>0.1, this is a temporary cut while a better cut being found.