Bug fix and added in a hostFlag into the catalogue
There was a bug fixed needed in the array which states if a halo had been already processed the orbits around it, in addition each halo now has a flag saying if it has been process in each orbital forestID to avoid duplicates.

The catalogue also contains a flag stating if the halo is a subhalo or not.