Bug fix to the creation of the orbital forest ID catalogues
There was a bug in the creation of the orbital forest ID catalogues, where orbits would be found around a branch even when it has merged with another branch causing dupilcations. Fixed an issure where, if a halo does not exist for a snapshot where the main branch terminates, the orbiting would continue to be tracked which causes an error since the main branch does not exist.