Update so halo now merge if the would of merged in Hydro sim
Currently if a halo goes within 0.1Rvir of its host it is said to of merged since in a hydrodynamical simulation the two central galaxies has merged. This is identified by a merged flag which shows after which point the galaxy has merged. The merger time scale will now be up to the time the galaxy has merged, if the galaxy has merged.