Added numSubStruct and RatioOfMassinSubsStruct datasets for halo and host
The amount of substructure that a halo has and the ratio of how much of its mass is in substructure is now outputted for both the orbiting halo and its host. This can be used to see if the orbiting halo or its host is a disturbed system leading to inconsistant results.