Update to distances are always physical
Before this a convert to comoving was done due to the periodicity correction needed in the interpolation routine, then it was converted back into physical in OrbWeaver as the data was being outputted. This has been removed so a conversion to comving is only done if interpolation is required so the periodicity correction can be done and then converted back just before the interpolation.