Added Eta to the output
The ratio of the (instantaneous) orbital angular momentum to the orbital angularmomentum of a circular orbit with the same orbital energy (Eta) is now outputted by OrbWeaver with J_circ being removed from the output. This is because Eta is more widely used and it is trivial to find J_circ from Eta (J_circ = L_orb/Eta).