orrection to the closest approach
Before this the closest approach was found using physical distances, but this would mean that most closest approaches happen at early snapshots where the physical distances are small. This has been corrected so the comparison is done in comovinging, but then converted back into physical distances.

So that the closest approach is also easily converted into comoving the scalefactor when it happens is also outputted.