Datasets from input catalog can be specified & update to preprocessing dataset names
The dataset that are to be loaded can be specified from the file in OrbitCatalogCreation/example_inputs/ depending on the input catalog used.

The name of the datasets that are outputted by the preprocessing have also been updated to leave them more ambigous since they are dependant on what the user desires as input.