Keeping the correct data when passage is deleted
Prior to this the closestapproach, closestapproachscalefactor and the minium Rscale and Rmax was set to be the values in the passage before, even though these quantities might of updated during the passages that are removed. This has now been updated so this information comes from the passage that is to be deleted.

The index of the passages is also beiing stored now for the orbit cleaning coming soon.