Implemented orbit cleaning
The cleaning is done to remove any "wobbles" in haloes orbits, where passages happen close together since the halo is orbiting/ interacting with another object. When a passage is removed the number of orbits in the crossing points needs to be updated. In addition, the orbit average quanties, num_entrytype and number of orbits needs to be updated for the passages following the deleted ones.

This cleaning cleans any orbits that falls within a region that is formed by ECCCLEANLIMIT and PHICLEANLIMIT. The parameters for cleaning where chosen by ensuring that the distribution of orbitecc_ratio matches that from Barber et al.,2014.